modle testing

+++ migrations/sqls/20220421220017-products-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products (
    id SERIAL PRIMARY KEY,
    name VARCHAR(50) NOT NULL,
-   price NUMERIC(3) NOT NULL,
+   price FLOAT(3) NOT NULL,
    stock INTEGER NOT NULL 
 );
